Add useSuggestions
