+++ packages/peregrine/lib/talons/SearchBar/index.js
@@ -1,3 +1,4 @@
 export { useAutocomplete } from './useAutocomplete';
 export { useSearchBar } from './useSearchBar';
 export { useSearchField } from './useSearchField';
+export { useSuggestions } from './useSuggestions';
