+++ packages/peregrine/lib/talons/SearchBar/useSearchField.js
@@ -1,3 +1,4 @@
+import { useCallback } from 'react';
 import { useFieldState, useFormApi } from 'informed';
 
 import { useSearchParam } from '../../hooks/useSearchParam';
